requirements.txt
torch==2.1.0
torchvision==0.16.0

beautifulsoup4==4.14.2
langdetect==1.0.9
lxml==6.0.2
marimo==0.17.6
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==10.4.0
pydantic==2.6.4
rfdetr==1.3.0
shapely==2.0.6
transformers==4.46.0
yattag==1.16.1

fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
